Create C.c
C. Welcome Everyone

In this pandemic situation it is hard to arrange a programming contest. But it is necessary to arrange a programming contest for the contestants or they become out of the practice and leave competitive programming. So this time BUBT Competitive Programming Community come to the front and arrange a programming contest on Hackerrank. In their first problem they welcome the contestants saying "Welcome to the BUBT IUPC". Then the contestants thanked the community by saying "Thank You" One judge has made a problem from this little conversation.
Input

You have to answer T test cases. In each test case You are given one Integer number(N) in a line which in 1 or 2
Output

In each line of test cases, print in a separate line "Welcome to the BUBT IUPC" without quotes if the given number is 1 or "Thank You" without quotes if the given number is 2
Samples