Create D.c
D. Tamim Iqbal - The One Man Army

    Problem Setter: Khairul Anam Mubin

    Dept. of CSE, Intake 38

"Tamim Iqbal" is a Bangladeshi left handed batsman who is called Mr. Dependable in his team. He often bat in opening and as he is dependable he tries to bat till the end of the match. He always uses his full strength and energy in a match.

In this Asia cup in Dubai, Bangladesh and India elected to play the final match but as "Tamim Iqbal" had injured and he can't bat so you and your friends are upset But you want to play a different mathematical game with your friends.

Given the overs of a match M and the balls in per overs N. You are to determine the maximum runs in a match "Tamim Iqbal" can take.

You know in opposition team India has good bowlers and fielders and they never give any no balls, wide balls and overthrow in fields.

Input

The overs of a match M and the balls per over N. Input is terminated by 'End of File'.

input limit (1 <= M,N <= 150).
Output

For every case of input there will be one line of output. It will contain the case number and the maximum runs "Tamim Iqbal" can take. Look at the sample output for exact formatting.
Samples
Input
Output