Create E.c
et there is a 2D black wheel. The difference between the circumference of outer side of the wheel and the circumference of inner side of the wheel is D meter. The ratio of the circumference of outer side of the wheel and the circumference of inner side of the wheel is K. You have to find the area of that black wheel.

See the figure below for understanding the wheel.



You have to find the area of that black shaded portion of the wheel. 

Input

The first line contains one integer T (1 ≤ T ≤ 105) — the number of test cases.

Each test case consists of two integers D and K where (1 <= D <= 10^3 , 2 <= K <= 10^3)
Output

For each test case, print the area in a single line. Error less than 1e-6 will be ignored.
Samples
Input